move mappers to new file
